+++ test/extensions/filters/network/postgres_proxy/sql/errors_and_warnings.sql
@@ -8,20 +8,20 @@ DELETE FROM test WHERE x=1;  -- (+) errors
 COMMIT;                      -- (+) statements, statements_other and transactions_rollback
 
 /*
-postgresql.egress_postgresql.errors: 1
-postgresql.egress_postgresql.frontend_commands: 6
-postgresql.egress_postgresql.sessions: 1
-postgresql.egress_postgresql.statements: 5
-postgresql.egress_postgresql.statements_delete: 0
-postgresql.egress_postgresql.statements_insert: 0
-postgresql.egress_postgresql.statements_other: 4
-postgresql.egress_postgresql.statements_select: 1
-postgresql.egress_postgresql.statements_update: 0
-postgresql.egress_postgresql.transactions: 1
-postgresql.egress_postgresql.transactions_commit: 0
-postgresql.egress_postgresql.transactions_rollback: 1
-postgresql.egress_postgresql.unrecognized: 0
-postgresql.egress_postgresql.warnings: 0
+postgres.egress_postgres.errors: 1
+postgres.egress_postgres.frontend_commands: 6
+postgres.egress_postgres.sessions: 1
+postgres.egress_postgres.statements: 5
+postgres.egress_postgres.statements_delete: 0
+postgres.egress_postgres.statements_insert: 0
+postgres.egress_postgres.statements_other: 4
+postgres.egress_postgres.statements_select: 1
+postgres.egress_postgres.statements_update: 0
+postgres.egress_postgres.transactions: 1
+postgres.egress_postgres.transactions_commit: 0
+postgres.egress_postgres.transactions_rollback: 1
+postgres.egress_postgres.unrecognized: 0
+postgres.egress_postgres.warnings: 0
 */
 
 /* WARNING:  there is already a transaction in progress */
@@ -31,20 +31,20 @@ SELECT 1;  -- (+) statements and statements_select
 COMMIT;    -- (+) statements, statements_other and transactions_commit
 
 /*
-postgresql.egress_postgresql.errors: 0
-postgresql.egress_postgresql.frontend_commands: 4
-postgresql.egress_postgresql.sessions: 1
-postgresql.egress_postgresql.statements: 4
-postgresql.egress_postgresql.statements_delete: 0
-postgresql.egress_postgresql.statements_insert: 0
-postgresql.egress_postgresql.statements_other: 3
-postgresql.egress_postgresql.statements_select: 1
-postgresql.egress_postgresql.statements_update: 0
-postgresql.egress_postgresql.transactions: 1
-postgresql.egress_postgresql.transactions_commit: 1
-postgresql.egress_postgresql.transactions_rollback: 0
-postgresql.egress_postgresql.unrecognized: 0
-postgresql.egress_postgresql.warnings: 1
+postgres.egress_postgres.errors: 0
+postgres.egress_postgres.frontend_commands: 4
+postgres.egress_postgres.sessions: 1
+postgres.egress_postgres.statements: 4
+postgres.egress_postgres.statements_delete: 0
+postgres.egress_postgres.statements_insert: 0
+postgres.egress_postgres.statements_other: 3
+postgres.egress_postgres.statements_select: 1
+postgres.egress_postgres.statements_update: 0
+postgres.egress_postgres.transactions: 1
+postgres.egress_postgres.transactions_commit: 1
+postgres.egress_postgres.transactions_rollback: 0
+postgres.egress_postgres.unrecognized: 0
+postgres.egress_postgres.warnings: 1
 */
 
 
@@ -54,18 +54,18 @@ SELECT 1+1;  -- (+) statements and statements_select
 ROLLBACK;    -- (+) statements, statements_other and transactions_rollback
 
 /*
-postgresql.egress_postgresql.errors: 0
-postgresql.egress_postgresql.frontend_commands: 3
-postgresql.egress_postgresql.sessions: 1
-postgresql.egress_postgresql.statements: 3
-postgresql.egress_postgresql.statements_delete: 0
-postgresql.egress_postgresql.statements_insert: 0
-postgresql.egress_postgresql.statements_other: 2
-postgresql.egress_postgresql.statements_select: 1
-postgresql.egress_postgresql.statements_update: 0
-postgresql.egress_postgresql.transactions: 1
-postgresql.egress_postgresql.transactions_commit: 0
-postgresql.egress_postgresql.transactions_rollback: 1
-postgresql.egress_postgresql.unrecognized: 0
-postgresql.egress_postgresql.warnings: 0
+postgres.egress_postgres.errors: 0
+postgres.egress_postgres.frontend_commands: 3
+postgres.egress_postgres.sessions: 1
+postgres.egress_postgres.statements: 3
+postgres.egress_postgres.statements_delete: 0
+postgres.egress_postgres.statements_insert: 0
+postgres.egress_postgres.statements_other: 2
+postgres.egress_postgres.statements_select: 1
+postgres.egress_postgres.statements_update: 0
+postgres.egress_postgres.transactions: 1
+postgres.egress_postgres.transactions_commit: 0
+postgres.egress_postgres.transactions_rollback: 1
+postgres.egress_postgres.unrecognized: 0
+postgres.egress_postgres.warnings: 0
 */